DSPHLE: Do the resampling correctly for PCM16 and AFC. Added linear interpolation (that works, this time :P)
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3764 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_DSP_HLE/Src/UCodes/UCode_Zelda.cpp
@@ -82,7 +82,7 @@ CUCode_Zelda::CUCode_Zelda(CMailHandler& _rMailHandler, u32 _CRC)
 	}
 
 	m_VoiceBuffer = new s32[256 * 1024];
-	m_ResampleBuffer = new s32[256 * 1024];
+	m_ResampleBuffer = new s16[256 * 1024];
 	m_LeftBuffer = new s32[256 * 1024];
 	m_RightBuffer = new s32[256 * 1024];
